+++ src/app/app.component.ts
@@ -1,5 +1,9 @@
 import {Component, Inject} from '@angular/core';
 import {AppConfig, APP_CONFIG} from './app.config';
+import {TinyTranslatorService} from './model/tiny-translator.service';
+import {isNullOrUndefined} from 'util';
+import {Observable} from 'rxjs/observable';
+import {Router} from '@angular/router';
 
 @Component({
   selector: 'app-root',
@@ -9,7 +13,7 @@ import {AppConfig, APP_CONFIG} from './app.config';
 export class AppComponent {
   title = 'app works!';
 
-  constructor(@Inject(APP_CONFIG) private APP_CONFIG: AppConfig) {
+  constructor(@Inject(APP_CONFIG) private APP_CONFIG: AppConfig, private translatorService: TinyTranslatorService, private router: Router) {
 
   }
 
@@ -21,5 +25,24 @@ export class AppComponent {
     return this.APP_CONFIG.BUILDVERSION;
   }
 
+  /**
+   * Auto translate all untranslated units.
+   * Redirects to a config page, if Google Translate is currently not available.
+   * Otherwise auto translates all untranslated units..
+   */
+  autoTranslate() {
+    this.translatorService.canAutoTranslate().subscribe((canTranslate: boolean) => {
+      if (canTranslate) {
+        this.translatorService.autoTranslate().subscribe((summary) => {
+          console.log('Summary: ', summary.content(), summary); // TODO show Toast or result page...
+        });
+      } else {
+        this.router.navigateByUrl('configureautotranslate');
+      }
+    });
+  }
 
+  configureAutoTranslate() {
+    this.router.navigateByUrl('configureautotranslate');
+  }
 }