+++ src/app/app.config.ts
@@ -7,8 +7,9 @@ import {environment} from '../environments/environment';
 
 export class AppConfig {
   // URL to Google translate API (optional)
-  GOOGLETRANSLATE_API_URL?: string;
-  // Your API Key
+  GOOGLETRANSLATE_API_ROOT_URL?: string;
+  // Your API Key, should not be set here, because it is secret
+  // can be typed in in the application, but test config needs it
   GOOGLETRANSLATE_API_KEY?: string;
   BUILDVERSION: string;
   BUILDTIME: string;
@@ -16,8 +17,10 @@ export class AppConfig {
 
 export const APP_CONFIG_VALUE: AppConfig = {
   // set values here
-  'BUILDVERSION': '0.2.0',
-  'BUILDTIME': '16.06.2017',
+  'BUILDVERSION': '0.3',
+  'BUILDTIME': '2017-07-01', // TODO should be dynamic
+  GOOGLETRANSLATE_API_ROOT_URL: 'https://translation.googleapis.com/',
+  GOOGLETRANSLATE_API_KEY: environment.googletranslate_api_key
 };
 
 export let APP_CONFIG = new InjectionToken<AppConfig>('app.config');