+++ src/app/app.routing.ts
@@ -5,6 +5,7 @@ import {TranslatePageComponent} from './translate-page/translate-page.component'
 import {ActiveProjectGuard} from './active-project.guard';
 import {CreateProjectPageComponent} from './create-project-page/create-project-page.component';
 import {EditProjectPageComponent} from './edit-project-page/edit-project-page.component';
+import {ConfigureAutoTranslatePageComponent} from './configure-auto-translate-page/configure-auto-translate-page.component';
 
 /**
  * Created by martin on 23.03.2017.
@@ -16,6 +17,7 @@ const appRoutes: Routes = [
   { path: 'createproject', component: CreateProjectPageComponent },
   { path: 'editproject', component: EditProjectPageComponent, canActivate: [ActiveProjectGuard] },
   { path: 'translate', component: TranslatePageComponent, canActivate: [ActiveProjectGuard] },
+  { path: 'configureautotranslate', component: ConfigureAutoTranslatePageComponent},
   { path: '', component: HomePageComponent}
 ];