+++ src/app/edit-project-page/edit-project-page.component.spec.ts
@@ -1,6 +1,8 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { EditProjectPageComponent } from './edit-project-page.component';
+import {ProjectEditorComponent} from '../project-editor/project-editor.component';
+import {AppModule} from '../app.module';
 
 describe('EditProjectPageComponent', () => {
   let component: EditProjectPageComponent;
@@ -8,7 +10,7 @@ describe('EditProjectPageComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ EditProjectPageComponent ]
+      imports: [AppModule]
     })
     .compileComponents();
   }));